requirements.txt
loguru==0.6.0
pyTelegramBotAPI==4.6.1
python-decouple==3.6
python-dotenv==0.20.0
python_telegram_bot_calendar==1.0.5
requests==2.26.0
